requirements.txt
django>=5.0.0,<5.1.0
channels>=3.0.0,<4.0.0
Pillow
djangorestframework